Fixed: Cast

+++ src/General.h
@@ -37,13 +37,13 @@ public:
     Ban(Window window, int color, double base_scale = 0.8, double scale = 0.85, int line_width = 2) {
         int windowMinSize = min(window.width, window.height);
         this->base_scale = base_scale;
-        this->base_width = windowMinSize * base_scale;
-        this->base_height = windowMinSize * base_scale;
+        this->base_width =  (double)windowMinSize * base_scale;
+        this->base_height = (double)windowMinSize * base_scale;
         this->base_ox = (window.width - this->base_width) / 2;
         this->base_oy = (window.height - this->base_height) / 2;
 
         this->scale = scale;
-        this->size = this->base_width * this->scale;
+        this->size = (int)(this->base_width * this->scale);
         this->ox = this->base_ox + (this->base_width - this->size) / 2;
         this->oy = this->base_oy + (this->base_height - this->size) / 2;
         this->masu_size = (this->size / 9);
